Refactor FPS tracking and scene initialization logic
Changed DebugContext frame storage from float to int32_t for FPS values and updated related calculations. Moved scene building order in Game::InitialiseAssemblers to ensure main menu is conditionally built before other scenes. Replaced Math::Lerp with Math::SmoothApproach for pointer update in MorseMonitorDisplaySystem. Minor window setup adjustments for improved configuration.

+++ include/core/debug_context.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "utility/morse_code.h"
+#include <cstdint>
 #include <deque>
 #include <string>
 
@@ -7,7 +8,7 @@
 struct DebugContext
 {
 	std::string receiverMessage = "";
-	std::deque<float> frames = std::deque<float>(32u, 0.0f);
+	std::deque<int32_t> frames = std::deque<int32_t>(32u, 0);
 	MorseCode::Pulse pulse = MorseCode::Invalid;
 	bool ignoreMainMenu = false;
 };
